+++ R/CCcolors-data.R
@@ -1,16 +1,18 @@
 #' @name CCcolors
 #' @aliases CCcolors
 #'
+#' @md
+#'
 #' @title Collaborative Cross colors
 #'
 #' @description A vector of 8 colors for use with the mouse
 #' Collaborative Cross and Diversity Outbreds.
 #'
 #' @details
-#' \code{CCorigcolors} are the original eight colors for the Collaborative Cross founder strains.
-#' \code{CCcolors} are slightly modified.
+#' `CCorigcolors` are the original eight colors for the Collaborative Cross founder strains.
+#' `CCcolors` are slightly modified.
 #'
-#' @source \url{http://compgen.unc.edu/wp/?page_id=577}
+#' @source <http://compgen.unc.edu/wp/?page_id=577>
 #'
 #' @keywords datasets
 #'